Ported old project and made a few necessary changes
The old project wasn't in the format that I think I want (I am still not
totally sure). Right now, to get things done by using this library,
things look downright ugly and aren't convenient (except in that the
code is largely done). I redid this for the purpose of making it easier
for the end user and I think I just made it more difficult.