Not working on new threads properly
The discover function actually runs on the same worker thread as
everything else, as far as I can tell. That might even be the main
thread (though I am not certain). I am implementing some threading, but
I don't know if it will work correctly, so we'll wait and see in this
branch off of main, to avoid totally trashing everything, of course.